% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Adlas.R
\name{Adlas}
\alias{Adlas}
\title{Sorted L1 parameter estimation solver for linear regression}
\usage{
Adlas(X, y, lambda, options = list())
}
\arguments{
\item{X}{A matrix}

\item{y}{A vector}

\item{lambda}{A vector}

\item{options}{An optional list}
}
\value{
The L1 parameter estimations for linear regression
}
\description{
Sorted L1 parameter estimation solver for linear regression
}
\examples{
Adlas(X, y, lambda)
}
