% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmSlope-package.R
\docType{package}
\name{glmSlope-package}
\alias{glmSlope}
\alias{glmSlope-package}
\title{glmSlope: What the Package Does in One 'Title Case' Line}
\description{
One paragraph description of what the package does
       as one or more full sentences.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/dkucharc/glmSlope}
  \item Report bugs at \url{https://github.com/dkucharc/glmSlope/issues}
}

}
\keyword{internal}
