% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LogisticSlope.R
\name{LogisticSlope}
\alias{LogisticSlope}
\title{Sorted L1 parameter estimation solver for logistic regression}
\usage{
LogisticSlope(X, y, lambda, options = list())
}
\arguments{
\item{X}{A matrix}

\item{y}{A vector}

\item{lambda}{A vector}

\item{options}{An optional list}
}
\value{
The L1 parameter estimations for logistic regression
}
\description{
Sorted L1 parameter estimation solver for logistic regression
}
