% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prox_sorted_L1.R
\name{prox_sorted_L1}
\alias{prox_sorted_L1}
\title{Compute the prox for the sorted L1 norm}
\usage{
prox_sorted_L1(x, lambda)
}
\arguments{
\item{x}{an input vector}

\item{lambda}{vectors of \eqn{\lambda}'s}
}
\value{
The prox estimations for the sorted L1 norm
}
\description{
Compute the prox for the sorted L1 norm
}
