% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proxSortedL1.R
\name{proxSortedL1}
\alias{proxSortedL1}
\title{Compute the prox for the sorted L1 norm}
\usage{
proxSortedL1(x, lambda)
}
\arguments{
\item{x}{an input vector}

\item{lambda}{vectors of $\lamda$'s}
}
\value{
The prox estimations for the sorted L1 norm
}
\description{
Compute the prox for the sorted L1 norm
}
\examples{
proxSortedL1(x, lambda)
}
